fix: recreate deleted stubs and fix all build errors across semantic, qdrant, gemmamemory, scripts packages

+++ internal/application/scripts/cache_eviction_usecase.go
@@ -161,7 +161,7 @@ func (u *CacheEvictionUseCase) Run(ctx context.Context, in CacheEvictionInput) (
 
 	return &CacheEvictionOutput{
 		CircuitBreakersReset: resetCount,
-		DeletedCount:         count,
+		DeletedCount:         int64(count),
 		EvictedTitles:        titles,
 	}, nil
 }
